requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
requests
# ecdsa
# paramiko
# pycrypto